finished users crud

+++ Homework 11/backend/src/v1/controllers/userController.mjs
@@ -4,6 +4,9 @@ import { validationResult } from "express-validator"
 
 class UserController {
   static async usersList(req, res) {
+    if (!req.user) {
+      return res.status(403).json({ success: false, errors: "Access denied" })
+    }
     try {
       const filters = {}
       for (const key in req.query) {
@@ -34,8 +37,13 @@ class UserController {
   }
 
   static async registerUser(req, res) {
+    if (!req.user) {
+      return res.status(403).json({ success: false, errors: "Access denied" })
+    }
     const errors = validationResult(req)
     const data = req.body
+    console.log(data)
+
     const types = await TypesDBService.getList()
 
     if (!errors.isEmpty()) {
@@ -49,15 +57,24 @@ class UserController {
     }
 
     try {
-      const { username, password, email } = req.body
-      const dataObj = { username, password, email }
+      const { username, password, confirmPassword, email } = req.body
+      const dataObj = { username, email }
 
       if (req.params.id) {
+        const user = await UsersDBService.getById(req.params.id)
+        const isMatch = await user.validPassword(confirmPassword)
+        if (!isMatch) {
+          return res.status(400).json({
+            errors: [{ msg: "Підтвердження пароля невдале" }],
+            data: dataObj,
+            types,
+            user: req.user,
+          })
+        }
         await UsersDBService.update(req.params.id, dataObj)
-      } else {
-        await UsersDBService.create(dataObj)
+        return res.status(200).json({ message: "User updated successfully" })
       }
-
+      await UsersDBService.create({ ...dataObj, password })
       res.status(200).json({ message: "User registered successfully" })
     } catch (err) {
       res.status(500).json({
@@ -71,8 +88,13 @@ class UserController {
 
   static async deleteUser(req, res) {
     try {
-      await UsersDBService.deleteById(req.body.id)
-      res.status(200).json({ success: true })
+      if (!req.user) {
+        return res.status(403).json({ success: false, errors: "Access denied" })
+      }
+      const user = await UsersDBService.deleteById(req.body.id)
+      const doLogout = user._id.equals(req.user._id)
+
+      res.status(200).json({ success: true, doLogout })
     } catch (error) {
       res.status(500).json({ success: false, message: "Failed to delete user" })
     }
